refactor(library): create calculator package
This diff creates a calculator package containing the
calculator proper and its config. This new package
is consistent with the structure of the iqb package,
where we use three levels of nesting.
