+++ library/src/iqb/calculator/__init__.py
@@ -1,0 +1,6 @@
+"""Package implementing IQB score calculation."""
+
+from .calculator import IQBCalculator
+from .config import IQB_CONFIG
+
+__all__ = ["IQB_CONFIG", "IQBCalculator"]
